significanlty improved script load strategy:
Requests: reduced the number of requests from 632 to 138. That's a reduction of approximately 78.2%.

Data Transferred: Reduced the data transferred from 4.2mb to 3.1mb. That's a reduction of approximately 26.2%.

Resources:Reduced the resources from 7.6mb to 3.1mb. That's a reduction of approximately 59.2%.

Finish Time: Reduced the finish time from 47.96s to 20.98s. That's a reduction of approximately 56.3%.

DOMContentLoaded: Reduced the DOMContentLoaded time from 291ms to 17ms. That's a reduction of approximately 94.2%.

Load Time: Reduced the load time from 359ms to 36ms. That's a reduction of approximately 90%.